#358 part1 move f08_brick to f09_brick
