+++ src/f09_brick/brick_config.json
