+++ src/f09_brick/brick_config.py
@@ -1,7 +1,7 @@
 from src.f00_instrument.file import open_file
 from src.f00_instrument.dict_tool import get_dict_from_json
 from src.f01_road.jaar_config import get_json_filename
-from src.f08_brick.examples.brick_env import src_brick_dir
+from src.f09_brick.examples.brick_env import src_brick_dir
 from os import getcwd as os_getcwd
 
 
@@ -10,7 +10,7 @@ def get_brick_config_file_name() -> str:
 
 
 def config_file_dir() -> str:
-    return f"{os_getcwd()}/src/f08_brick"
+    return f"{os_getcwd()}/src/f09_brick"
 
 
 def get_brick_config_dict() -> dict:
