+++ src/f09_brick/brick_formats/brick_format_00036_problem_healer_v0_0_0.json
