+++ src/f09_brick/brick_formats/brick_format_00040_bridge_otx_to_inx_v0_0_0.json
