+++ src/f09_brick/brick_formats/brick_format_00041_bridge_explicit_label_v0_0_0.json
