+++ src/f09_brick/test_brick_config/test_brick_config_.py
@@ -32,7 +32,7 @@ from src.f07_fiscal.fiscal_config import (
     fiscal_timeline_month_str,
     fiscal_timeline_weekday_str,
 )
-from src.f08_brick.brick_config import (
+from src.f09_brick.brick_config import (
     time_id_str,
     brick_number_str,
     allowed_crud_str,
@@ -46,7 +46,7 @@ from src.f08_brick.brick_config import (
     brick_format_00020_bud_acct_membership_v0_0_0,
     brick_format_00013_itemunit_v0_0_0,
 )
-from src.f08_brick.filter_config import get_filter_categorys
+from src.f08__filter.filter_config import get_filter_categorys
 from os import getcwd as os_getcwd
 
 
@@ -56,7 +56,7 @@ def test_get_brick_config_file_name_ReturnsObj():
 
 
 def test_config_file_dir_ReturnsObj() -> str:
-    assert config_file_dir() == f"{os_getcwd()}/src/f08_brick"
+    assert config_file_dir() == f"{os_getcwd()}/src/f09_brick"
 
 
 def test_get_brick_config_dict_ReturnsObj():
