+++ src/f09_brick/test_brick_format/test_env.py
@@ -1,4 +1,4 @@
-from src.f08_brick.examples.brick_env import (
+from src.f09_brick.examples.brick_env import (
     brick_examples_dir,
     src_brick_examples_dir,
     brick_fiscals_dir,
@@ -7,7 +7,7 @@ from src.f08_brick.examples.brick_env import (
 
 def test_src_brick_examples_dir_ReturnsObj():
     # ESTABLISH / WHEN / THEN
-    assert src_brick_examples_dir() == "src/f08_brick/examples"
+    assert src_brick_examples_dir() == "src/f09_brick/examples"
 
 
 def test_brick_examples_dir_ReturnsObj():
