+++ src/f10_world/world_tool.py
@@ -1,5 +1,5 @@
-from src.f08_brick.brick_config import get_brick_numbers
-from src.f08_brick.pandas_tool import get_all_excel_sheetnames
+from src.f09_brick.brick_config import get_brick_numbers
+from src.f09_brick.pandas_tool import get_all_excel_sheetnames
 
 
 def get_all_excel_bricksheets(dir: str) -> set[str, str, str]:
